Changed makeRequestDownloadResponse to directly pass the memory stream instance, since that class has some useful features. Modified UpdateController to use an ASN1 based collection of data instead of a binary writer and moved the processing of if there are updates to the client. Fixed some int plugin ids to be longs. Changed readServerPluginInfo to use an ASN1 based format and to do the processing on the client.
git-svn-id: https://192.168.3.139:8443/svn/Medical/trunk@3957 956a8c8d-885e-cb46-a0b9-a2f355b4c364

+++ AnomalousMedical/GUI/DownloadManagerGUI/ServerPluginDownloadInfo.cs
@@ -13,7 +13,7 @@ namespace Medical.GUI
         private String description = null;
         private bool readingDescriptionFromServer = false;
 
-        public ServerPluginDownloadInfo(DownloadManagerServer server, int pluginId, String name, ServerDownloadStatus status)
+        public ServerPluginDownloadInfo(DownloadManagerServer server, long pluginId, String name, ServerDownloadStatus status)
             :base(status)
         {
             this.server = server;
@@ -55,7 +55,7 @@ namespace Medical.GUI
             return new UpdateInfo(ImageKey, Name, String.Format("You must restart Anomalous Medical to finish installing {0}. You may install more things by selecting them and clicking install before restarting.", Name), ServerDownloadStatus.PendingInstall, false);
         }
 
-        public int PluginId { get; set; }
+        public long PluginId { get; set; }
 
         public override string MoreInfoURL
         {
